Rename build params function
