+++ src/app/components/ATIAnalytics/params/buildParams/index.test.ts
@@ -1,8 +1,8 @@
 import { TOPIC_PAGE } from '#app/routes/utils/pageTypes';
 import * as analyticsUtils from '#lib/analyticsUtils';
-import { RequestContextProps } from '../../../contexts/RequestContext';
-import { ServiceConfig } from '../../../models/types/serviceConfig';
-import { buildPageATIParams, buildPageATIUrl } from './buildParams';
+import { RequestContextProps } from '../../../../contexts/RequestContext';
+import { ServiceConfig } from '../../../../models/types/serviceConfig';
+import { buildPageATIParams, buildPageATIUrl } from '.';
 
 (analyticsUtils.getAtUserId as jest.Mock) = jest.fn();
 (analyticsUtils.getCurrentTime as jest.Mock) = jest
